abc144/src/test/java/abc144/e/MainTest.java
package abc144.e;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.PrintStream;
import org.junit.Test;

public class MainTest {

    @Test
    public void test1() {
        String input = "3 5\n"
            + "4 2 1\n"
            + "2 3 1";
        String expected = "2";

        assertThat(execute(input + "\n"), is(expected + "\n"));
    }

    @Test
    public void test2() {
        String input = "3 8\n"
            + "4 2 1\n"
            + "2 3 1";
        String expected = "0";

        assertThat(execute(input + "\n"), is(expected + "\n"));
    }

    @Test
    public void test3() {
        String input = "11 14\n"
            + "3 1 4 1 5 9 2 6 5 3 5\n"
            + "8 9 7 9 3 2 3 8 4 6 2";
        String expected = "12";

        assertThat(execute(input + "\n"), is(expected + "\n"));
    }

    private String execute(String input) {
        ByteArrayOutputStream os = new ByteArrayOutputStream();
        Main.solve(
                new ByteArrayInputStream(input.getBytes()), new PrintStream(os)
        );
        return os.toString().replace(System.lineSeparator(), "\n");
    }
}
